[FEAT] Backported fastcopy.c from c-blosc1
